refactor out tray context menu from sidebar
